fix: aslo hide salary information in experience detail
